refactor : change racingcar.* to MVC pattern
