+++ src/main/java/racingcar/view/Input.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.domain.InvalidInputException;
 
 import java.util.Arrays;
 import java.util.HashSet;
